requirements.txt
python-binance==1.0.19
websocket-client==1.8.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
prometheus-client==0.20.0

scikit-learn==1.5.1
xgboost==2.0.3

requests==2.32.3
tweepy==4.14.0
textblob==0.18.0.post0

pytest==8.3.2
pytest-asyncio==0.23.8
mypy==1.10.0
ruff==0.5.7
